# docs/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: AgentCrew Sidecar API
  version: 0.5.0-draft
  description: |
    单用户本地 sidecar。M0 操作带完整请求/响应/错误 schema；
    M2 操作定义当前身份、资源、授权、审计和诊断边界；实施归属见 x-domain-card。
    演示身份通过服务端签发的 X-AgentCrew-Identity 标识固定于每个请求。
    p3/p5 操作 schema 随对应切片细化（当前为方向性描述）。
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port: { default: "8710" }
security:
  - bearerToken: []
  - bearerToken: []
    demoIdentity: []
...
components:
  securitySchemes:
    bearerToken: { type: http, scheme: bearer }
    demoIdentity: { type: apiKey, in: header, name: X-AgentCrew-Identity, description: 服务端签发的演示身份标识，Bearer仍然必需 }
  schemas:
...
    Identity:
      type: object
      required: [credential_owner_id, effective_user_id, name, org_id, role, demo, workspace_ids]
      properties:
        credential_owner_id: { type: string }
        effective_user_id: { type: string }
        name: { type: string }
        org_id: { type: string }
        role: { type: string, enum: [owner, admin, member] }
        demo: { type: boolean }
        workspace_ids: { type: array, items: { type: string } }
...
    AuditRecord:
      type: object
      required: [seq, ts, actor_type, actor_id, action, resource_type, resource_id, detail, prev_hash, hash]
      properties:
        seq: { type: integer, minimum: 1 }
        ts: { type: string, format: date-time }
        actor_type: { type: string, enum: [user, agent, system, curator] }
        actor_id: { type: string }
        action: { type: string }
        resource_type: { type: [string, "null"] }
        resource_id: { type: [string, "null"] }
        detail: { type: object, additionalProperties: true, description: 已脱敏，包含真实及有效身份和范围 }
        prev_hash: { type: string }
        hash: { type: string }
    AuditVerification:
      type: object
      required: [ok, internal, anchor]
      properties:
        ok: { type: boolean }
        internal:
          type: object
          required: [ok, checked_rows, broken_at, reason]
          properties:
            ok: { type: boolean }
            checked_rows: { type: integer, minimum: 0 }
            broken_at: { type: [integer, "null"] }
            reason: { type: [string, "null"] }
        anchor:
          type: object
          required: [ok, status, seq, hash]
          properties:
            ok: { type: boolean }
            status: { type: string, enum: [verified, empty, missing, malformed, mismatch] }
            seq: { type: [integer, "null"] }
            hash: { type: [string, "null"] }
    DiagnosticBackup:
      type: object
      required: [id, kind, created_at, verified, database_sha256, anchor_seq, anchor_hash]
      properties:
        id: { type: string }
        kind: { type: string, enum: [database, directory] }
        created_at: { type: string, format: date-time }
        verified: { type: boolean }
        database_sha256: { type: string }
        anchor_seq: { type: integer }
        anchor_hash: { type: string }
  parameters:
...
    GovernanceWorkspace: { name: workspace_id, in: query, schema: { type: string, minLength: 1 } }
    GovernanceLimit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    GovernanceAfter: { name: after, in: query, schema: { type: string, minLength: 1 }, description: 排他游标绑定有效身份、范围、过滤及稳定排序 }
...
paths:
  /api/identity:
    get:
      operationId: readIdentity
      tags: [governance]
      x-implementation-card: M2-03
      x-domain-card: M2-03
      responses:
        <<: *memory_errors
        "200":
          description: 请求固定的有效身份及当前角色
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Identity" }
...
  /api/audit:
    get:
      operationId: listAudit
      tags: [audit]
      x-implementation-card: M2-10
      x-domain-card: M2-10
      parameters: &audit_parameters
        - { $ref: "#/components/parameters/GovernanceWorkspace" }
        - { $ref: "#/components/parameters/GovernanceLimit" }
        - { $ref: "#/components/parameters/GovernanceAfter" }
        - { name: actor, in: query, schema: { type: string } }
        - { name: action, in: query, schema: { type: string } }
        - { name: resource, in: query, schema: { type: string }, description: resource_type或resource_type:resource_id }
      responses:
        <<: *memory_errors
        "200": &audit_page_response
          description: 脱敏审计记录按seq降序；member仅自身actor
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [items, next_after]
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/AuditRecord" } }
                      next_after: { type: [string, "null"] }
  /api/audit/export:
    get:
      operationId: exportAudit
      tags: [audit]
      x-implementation-card: M2-10
      x-domain-card: M2-10
      parameters: *audit_parameters
      responses:
        <<: *memory_errors
        "200": *audit_page_response
  /api/audit/verify:
    post:
      operationId: verifyAudit
      tags: [audit]
      x-implementation-card: M2-10
      x-domain-card: M2-10
      responses:
        <<: *memory_errors
        "200":
          description: owner执行两级校验；失败进入只读诊断，损坏链禁止追加验证审计
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/AuditVerification" }
  /api/diagnostics/backups:
    get:
      operationId: listDiagnosticBackups
      tags: [audit]
      x-implementation-card: M2-10
      x-domain-card: M2-10
      responses:
        <<: *memory_errors
        "200":
          description: 服务管理的备份；恢复前重新验证SHA及两级审计
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/DiagnosticBackup" } }
    post:
      operationId: createDiagnosticBackup
      tags: [audit]
      x-implementation-card: M2-10
      x-domain-card: M2-10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [change_id, kind]
              properties:
                change_id: { type: string, minLength: 1 }
                kind: { type: string, enum: [database, directory] }
      responses:
        <<: *memory_errors
        "200":
          description: 正常模式创建并验证备份；诊断期间禁止创建
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/DiagnosticBackup" }
  /api/diagnostics/restore:
    post:
      operationId: restoreDiagnosticBackup
      tags: [audit]
      x-implementation-card: M2-10
      x-domain-card: M2-10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [backup_id, change_id]
              properties:
                backup_id: { type: string, minLength: 1 }
                change_id: { type: string, minLength: 1 }
      responses:
        <<: *memory_errors
        "200":
          description: 受控恢复保留损坏材料；重启并通过完整验证后才能恢复业务
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [backup_id, restart_required, preserved_path]
                    properties:
                      backup_id: { type: string }
                      restart_required: { type: boolean, const: true }
                      preserved_path: { type: string }
...